Create UI for rock-paper-scissors game in rps-ui
DOM manipulation and events ultilized to change the
game from being played in the console to having a UI
consisting of choice buttons and div's to display
the running score and corresponding results text.
An event listener for the buttons calls playRound()
which was updated to show each result in the DOM;
winning text is shown once one player has 5 points.